Fix watching for geofence changes
