+++ src/weatherWorker.js
@@ -92,7 +92,7 @@ function reloadDts() {
 
 function reloadGeofence() {
 	try {
-		const newGeofence = require('./lib/geofenceLoader').readGeofenceFile(config, path.join(__dirname, `../${config.geofence.path}`))
+		const newGeofence = require('./lib/geofenceLoader').readAllGeofenceFiles(config)
 		weatherController.setGeofence(newGeofence)
 		log.info('Geofence reloaded')
 	} catch (err) {
